requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.2
kornia==0.6.3
Markdown==3.3.6
multidict==6.0.2
munch==2.5.0
numpy==1.22.2
oauthlib==3.2.0
packaging==21.3
pathtools==0.1.2
Pillow==9.0.1
pip==20.0.2
pkg-resources==0.0.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytorch-lightning==1.5.10
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
segmentation-models-pytorch==0.3.0.dev0
sentry-sdk==1.5.7
setproctitle==1.2.2
setuptools==59.5.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
timm==0.4.12
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchmetrics==0.7.2
torchvision==0.11.3+cu113
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
wandb==0.12.11
Werkzeug==2.0.3
wheel==0.37.1
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0